requirements.txt
fastapi
uvicorn
langchain
langchain-groq
langgraph
pydantic
python-dotenv